+++ src/components/JSONTree/JSONTree.scss
@@ -1,6 +1,6 @@
 @use '../../styles/mixins.scss';
 
-.ydb-json-tree-with-search {
+.ydb-json-tree {
     position: relative;
 
     width: 100%;
