+++ src/components/JSONTree/i18n/en.json
