+++ src/components/JSONTree/i18n/index.ts
@@ -2,6 +2,6 @@ import {registerKeysets} from '../../../utils/i18n';
 
 import en from './en.json';
 
-const COMPONENT = 'ydb-json-tree-with-search';
+const COMPONENT = 'ydb-json-tree';
 
 export default registerKeysets(COMPONENT, {en});
